LayoutTests/imported/w3c/web-platform-tests/fetch/api/credentials/authentication-basic-expected.txt

localhost:8800 - didReceiveAuthenticationChallenge - Simulating cancelled authentication sheet
localhost:8800 - didReceiveAuthenticationChallenge - Simulating cancelled authentication sheet
localhost:8800 - didReceiveAuthenticationChallenge - Simulating cancelled authentication sheet

FAIL User-added Authorization header with include mode assert_equals: HTTP status is 200 expected 200 but got 401
FAIL User-added Authorization header with same-origin mode assert_equals: HTTP status is 200 expected 200 but got 401
FAIL User-added Authorization header with omit mode assert_equals: HTTP status is 200 expected 200 but got 401
